split Hash out into separate hash.rs file in pbc lib
