+++ crypto/src/pbc/fast.rs
@@ -41,6 +41,7 @@
 // --------------------------------------------------------------------------
 
 use super::*;
+use self::hash::*;
 
 #[derive(Copy, Clone)]
 pub struct Zr([u8;ZR_SIZE_AR160]);
