+++ crypto/src/pbc/mod.rs
@@ -45,6 +45,7 @@ use rust_libpbc;
 
 pub mod secure;
 pub mod fast;
+pub mod hash;
 
 // -------------------------------------------------------------------
 // Fast AR160 curves, but low security 2^80
@@ -246,66 +247,4 @@ pub fn init_pairings() {
         println!("");
     }
 }
-
-// -----------------------------------------------------
-// Hashing with SHA3
-
-const HASH_SIZE : usize = 32;
-
-#[derive(Copy, Clone)]
-pub struct Hash([u8; HASH_SIZE]);
-
-impl Hash {
-    pub fn base_vector(&self) -> &[u8] {
-        &self.0
-    }
-
-    pub fn from_vector(msg : &[u8]) -> Hash {
-        hash(msg)
-    }
-
-    pub fn to_str(&self) -> String {
-        u8v_to_typed_str("H", &self.base_vector())
-    }
-}
-
-pub fn hash(msg : &[u8]) -> Hash {
-    let mut hasher = Sha3_256::new();
-    hasher.input(msg);
-    let out = hasher.result();
-    let mut h = [0u8; HASH_SIZE];
-    h.copy_from_slice(&out[.. HASH_SIZE]);
-    Hash(h)
-}
-
-pub fn hash_nbytes(nb : usize, msg : &[u8]) -> Vec<u8> {
-    let nmsg = msg.len();
-    let mut ct = nb;
-    let mut ans = vec![0u8; nb];
-    let mut jx = 0;
-    let mut kx = 0u8;
-    while ct > 0 {
-        let mut inp = vec![kx];
-        for ix in 0 .. nmsg {
-            inp.push(msg[ix]);
-        }
-        let mut hasher = Sha3_256::new();
-        hasher.input(inp);
-        let out = hasher.result();
-        let end = if ct > HASH_SIZE {
-                    HASH_SIZE
-                } else {
-                    ct
-                };
-        for ix in 0 .. end {
-            ans[jx + ix] = out[ix];
-        }
-        jx += end;
-        ct -= end;
-        kx += 1;
-    }
-    ans
-}
-
-
