+++ crypto/src/pbc/secure.rs
@@ -33,6 +33,7 @@
 // --------------------------------------------------------------------------
 
 use super::*;
+use self::hash::*;
 
 #[derive(Copy, Clone)]
 pub struct Zr([u8;ZR_SIZE_FR256]);
