+++ examples/pbc.rs
@@ -28,6 +28,7 @@ extern crate stegos_crypto;
 extern crate rust_libpbc;
 
 use stegos_crypto::pbc::*;
+use stegos_crypto::pbc::hash::*;
 
 use std::sync::Mutex;
